Support phrases in arrays:
Issue:
By default, passing in an array to the `phrases:` attribute will result
in the following translation phrases:
- `items.0`
- `items.1`
- ...

This works fairly well if you know how many entries are in the array
provided, but for my personal use it would be more convenient to
receive an array from the `items` translation.

Solution:
`items` now returns an array of transformed phrases in the same order as
retrieved from the `phrases:` translations object.